requirements.txt
line-bot-sdk
streamlink
opencv-python
Flask
pyyaml
argparse
google-api-python-client
google-auth-httplib2
google-auth-oauthlib